+++ packages/create/test/unit/cli/add/mockserver-config.test.ts
@@ -5,7 +5,7 @@ import { addAddMockserverConfigCommand } from '../../../../src/cli/add/mockserve
 import * as prompts from 'prompts';
 import * as mockserverWriter from '@sap-ux/mockserver-config-writer';
 import * as logger from '../../../../src/tracing/logger';
-import * as childProcess from 'child_process';
+import * as childProcess from 'node:child_process';
 import * as npmCommand from '@sap-ux/project-access';
 import { join } from 'node:path';