+++ packages/fiori-app-sub-generator/src/utils/command-runner.ts
@@ -1,7 +1,7 @@
 import type { ILogWrapper } from '@sap-ux/fiori-generator-shared';
 import { t } from './i18n';
-import { spawn } from 'child_process';
-import type { SpawnOptionsWithoutStdio } from 'child_process';
+import { spawn } from 'node:child_process';
+import type { SpawnOptionsWithoutStdio } from 'node:child_process';
 
 /**
  *