+++ packages/fiori-app-sub-generator/test/unit/fiori-app-generator/install.test.ts
@@ -1,6 +1,6 @@
 import type { CapService } from '@sap-ux/cap-config-writer';
 import { DefaultLogger } from '@sap-ux/fiori-generator-shared';
-import childProcess from 'child_process';
+import childProcess from 'node:child_process';
 import os from 'node:os';
 import { installDependencies } from '../../../src/fiori-app-generator/install';
 import { CommandRunner, initI18nFioriAppSubGenerator, t } from '../../../src/utils';