+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.73
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/fe-fpm-writer@0.40.1
+
 ## 0.1.72
 
 ### Patch Changes