+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/cap-config-writer
 
+## 0.12.36
+
+### Patch Changes
+
+- Updated dependencies [a9471d0]
+    - @sap-ux/project-access@1.32.17
+    - @sap-ux/logger@0.7.3
+    - @sap-ux/fiori-generator-shared@0.13.45
+
 ## 0.12.35
 
 ### Patch Changes