+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 2.5.17
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/ui5-application-writer@1.6.6
+    - @sap-ux/odata-service-writer@0.29.1
+    - @sap-ux/ui5-test-writer@0.7.45
+    - @sap-ux/ui5-config@0.29.12
+    - @sap-ux/logger@0.7.3
+    - @sap-ux/fiori-generator-shared@0.13.45
+    - @sap-ux/cap-config-writer@0.12.36
+
 ## 2.5.16
 
 ### Patch Changes