+++ packages/flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/flp-config-sub-generator
 
+## 0.3.88
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/flp-config-inquirer@0.4.85
+    - @sap-ux/app-config-writer@0.6.70
+    - @sap-ux/inquirer-common@0.10.2
+    - @sap-ux/project-access@1.32.17
+    - @sap-ux/i18n@0.3.7
+    - @sap-ux/deploy-config-generator-shared@0.1.64
+    - @sap-ux/fiori-generator-shared@0.13.45
+
 ## 0.3.87
 
 ### Patch Changes