+++ packages/preview-middleware-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux-private/preview-middleware-client
 
+## 0.18.7
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+    - @sap-ux-private/control-property-editor-common@0.7.3
+
 ## 0.18.6
 
 ### Patch Changes