+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/telemetry
 
+## 0.6.49
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/project-access@1.32.17
+    - @sap-ux/nodejs-utils@0.2.11
+    - @sap-ux/ui5-config@0.29.12
+    - @sap-ux/logger@0.7.3
+    - @sap-ux/store@1.4.1
+    - @sap-ux/btp-utils@1.1.6
+
 ## 0.6.48
 
 ### Patch Changes